+++ Source/DSP/ProPhatSynthesiser.cpp
