+++ Source/DSP/ProPhatVoice.h
@@ -19,7 +19,7 @@
 #pragma once
 //#include "UI/ButtonGroupComponent.h"
 #include "GainedOscillator.h"
-#include "Helpers.h"
+#include "../Utility/Helpers.h"
 
 #include <mutex>
 #include <set>
