+++ Source/UI/ButtonGroupComponent.h
@@ -17,7 +17,7 @@
 */
 #pragma once
 
-#include "../Helpers.h"
+#include "../Utility/Helpers.h"
 #include <memory>
 
 enum defaults
