+++ Source/Utility/Helpers.h
@@ -19,18 +19,6 @@
 #pragma once
 #include <JuceHeader.h>
 
-#ifndef CPU_USAGE
-    #define CPU_USAGE 0
-#endif
-
-#ifndef DEBUG_VOICES
-    #define DEBUG_VOICES 0
-#endif
-
-#ifndef PRINT_ALL_SAMPLES
-    #define PRINT_ALL_SAMPLES 0
-#endif
-
 namespace Constants
 {
 constexpr auto groupComponentFontHeight { 24 };
