+++ Source/Utility/Macros.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#ifndef CPU_USAGE
+#define CPU_USAGE 0
+#endif
+
+#ifndef DEBUG_VOICES
+#define DEBUG_VOICES 0
+#endif
+
+#ifndef PRINT_ALL_SAMPLES
+#define PRINT_ALL_SAMPLES 0
+#endif
+
+#ifndef USE_NATIVE_TITLE_BAR
+#define USE_NATIVE_TITLE_BAR 0
+#endif
